pyproject.toml
[project]
name = "googlee"
version = "0.1.0"
description = "Hackathon"
requires-python = ">=3.10"

dependencies = [
    "google-adk>=1.22.0",
    "pillow>=12.1.0",
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "python-dotenv>=1.0.0",
]

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]

